requirements.txt
# Core libraries
torch
torchvision
transformers
numpy
Pillow
pyyaml
tqdm

# FAISS (CPU). If you have CUDA and want GPU support, install `faiss-gpu` instead.
faiss-cpu

# Optional but commonly required for some tokenizers
sentencepiece

# Optional helpers
accelerate

# Description optimization dependencies
sentence-transformers
hdbscan
scikit-learn
peft
datasets
evaluate
mmdet
mmcv
